Work on query cache

+++ cache/engines/WaxCacheFilesystem.php
@@ -18,14 +18,9 @@ class WaxCacheFilesystem implements CacheEngine{
   public function __construct($key, $options) {
     foreach($options as $k=>$option) $this->$k = $option;
     if(!$this->root_dir) $this->root_dir = CACHE_DIR;
-    if(strpos($key, "/")) {
-      $file = substr(strrchr($key, "/"),1);
-      $path = strrpos($key, "/") ? substr($key, 0, strrpos($key, "/")+1) : false;
-      $this->key = $file;
-    } else $this->key=$key;
-    $this->root_dir .= $path;
-    if(!is_writable($this->root_dir)) mkdir($this->root_dir,0777,true);
-    if(!is_writable($this->root_dir)) $this->writable = false;
+    $this->key = $key;
+    if(!is_writable($this->file())) mkdir($this->file(),0777,true);
+    if(!is_writable($this->file())) $this->writable = false;
     else $this->writable = true;
   }
 	
@@ -48,13 +43,27 @@ class WaxCacheFilesystem implements CacheEngine{
     
 	}
 	
-	public function expire() {
-    return unlink($this->key());
+	public function expire($query = false) {
+    if(!$query) return unlink($this->key());
+    else $this->flush($query);
 	}
 	
 	public function key() {
+	  return $this->key_path($this->key).$this->key;
+	}
+	
+	public function file() {
 	  return $this->root_dir.$this->key;
 	}
+	
+	public function flush($query) {
+	  if(strlen($query)>1) {
+	    if(is_dir($this->root_dir.$query)) {
+	      File::recursively_delete($this->root_dir.$query);
+	    }
+	  } else foreach(scandir($this->root_dir) as $dir) File::recursively_delete($this->root_dir.$dir);
+	}
+	
 
 }
